Rename particles.isotopes
