+++ plasmapy/particles/_isotopes.py
