+++ plasmapy/particles/tests/test_particle_class.py
@@ -10,6 +10,7 @@ from astropy import units as u
 from astropy.constants import c, e, m_e, m_n, m_p
 
 from plasmapy.particles import json_load_particle, json_loads_particle, molecule
+from plasmapy.particles._isotopes import data_about_isotopes
 from plasmapy.particles.atomic import known_isotopes
 from plasmapy.particles.exceptions import (
     ChargeError,
@@ -22,7 +23,6 @@ from plasmapy.particles.exceptions import (
     ParticleError,
     ParticleWarning,
 )
-from plasmapy.particles.isotopes import data_about_isotopes
 from plasmapy.particles.particle_class import (
     CustomParticle,
     DimensionlessParticle,
